# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: NearBuy Business API (MongoDB-backed)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shops/{shopId}/products:
    get:
      operationId: listProducts
      tags: [products]
      parameters:
        - $ref: "#/components/parameters/ShopId"
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
    post:
      operationId: createProduct
      tags: [products]
      parameters:
        - $ref: "#/components/parameters/ShopId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductCreateInput" }
      responses:
        "200":
          description: Created product
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
  /shops/{shopId}/products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - $ref: "#/components/parameters/ShopId"
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductUpdateInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - $ref: "#/components/parameters/ShopId"
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeleteResult" }
  /shops/{shopId}/products/analyze-photo:
    post:
      operationId: analyzeProductPhoto
      tags: [products]
      parameters:
        - $ref: "#/components/parameters/ShopId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnalyzePhotoInput" }
      responses:
        "200":
          description: AI suggestions
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductSuggestion" }
  /shops/{shopId}/products/{productId}/discounts:
    get:
      operationId: listDiscounts
      tags: [discounts]
      parameters:
        - $ref: "#/components/parameters/ShopId"
        - in: path
          name: productId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Discount" }
    post:
      operationId: createDiscount
      tags: [discounts]
      parameters:
        - $ref: "#/components/parameters/ShopId"
        - in: path
          name: productId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DiscountCreateInput" }
      responses:
        "200":
          description: Created discount
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Discount" }
  /shops/{shopId}/products/{productId}/discounts/{discountId}:
    delete:
      operationId: deleteDiscount
      tags: [discounts]
      parameters:
        - $ref: "#/components/parameters/ShopId"
        - in: path
          name: productId
          required: true
          schema: { type: string }
        - in: path
          name: discountId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeleteResult" }
...
  /shops/{shopId}/members:
    get:
      operationId: listShopMembers
      tags: [members]
      parameters:
        - $ref: "#/components/parameters/ShopId"
      responses:
        "200":
          description: Members and pending invitations
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShopMembersList" }
    post:
      operationId: inviteShopMember
      tags: [members]
      parameters:
        - $ref: "#/components/parameters/ShopId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InviteMemberInput" }
      responses:
        "200":
          description: Created invitation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShopInvitation" }
  /shops/{shopId}/members/{userId}:
    delete:
      operationId: removeShopMember
      tags: [members]
      parameters:
        - $ref: "#/components/parameters/ShopId"
        - in: path
          name: userId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeleteResult" }
  /shops/{shopId}/invitations/{invitationId}:
    delete:
      operationId: cancelShopInvitation
      tags: [members]
      parameters:
        - $ref: "#/components/parameters/ShopId"
        - in: path
          name: invitationId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeleteResult" }
...
  /invitations/{token}/accept:
    post:
      operationId: acceptInvitation
      tags: [members]
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Accepted; returns shop with role
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShopWithRole" }
...
components:
  parameters:
    ShopId:
      in: path
      name: shopId
      required: true
      schema: { type: string }
  schemas:
...
    Shop:
      type: object
      properties:
        id: { type: string }
        sellerId: { type: string }
        name: { type: string }
        marketName: { type: string, nullable: true }
        stallInfo: { type: string, nullable: true }
        latitude: { type: number }
        longitude: { type: number }
        isOpen: { type: boolean }
      required: [id, sellerId, name, latitude, longitude, isOpen]
    ShopWithRole:
      type: object
      properties:
        shop: { $ref: "#/components/schemas/Shop" }
        role:
          type: string
          enum: [seller, sub_seller]
      required: [shop, role]
...
    Category:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        parent: { type: string, nullable: true }
        icon: { type: string, nullable: true }
      required: [id, name, slug]
    Dimension:
      type: object
      properties:
        height: { type: number, minimum: 0 }
        length: { type: number, minimum: 0 }
        width: { type: number, minimum: 0 }
      required: [height, length, width]
    ProductVariation:
      type: object
      properties:
        id: { type: string, nullable: true }
        sku: { type: string }
        price: { type: number, minimum: 0 }
        quantity: { type: number, minimum: 0 }
        colors:
          type: array
          items: { type: string }
        photos:
          type: array
          items: { type: string }
        dimension:
          oneOf:
            - { $ref: "#/components/schemas/Dimension" }
            - { type: "null" }
      required: [sku, price, quantity, colors, photos]
    Product:
      type: object
      properties:
        id: { type: string }
        shopId: { type: string }
        sellerId: { type: string }
        name: { type: string }
        slug: { type: string, nullable: true }
        brand: { type: string, nullable: true }
        description: { type: string, nullable: true }
        price: { type: number }
        quantity: { type: number }
        colors:
          type: array
          items: { type: string }
        photos:
          type: array
          items: { type: string }
        sizes:
          type: array
          items: { type: string }
        categories:
          type: array
          items: { $ref: "#/components/schemas/Category" }
        weight: { type: number, nullable: true }
        dimension:
          oneOf:
            - { $ref: "#/components/schemas/Dimension" }
            - { type: "null" }
        variations:
          type: array
          items: { $ref: "#/components/schemas/ProductVariation" }
        rating: { type: number, nullable: true }
        reviews: { type: number }
        totalSell: { type: number }
        applyDiscount: { type: boolean }
        tags:
          type: array
          items: { type: string }
        stockStatus:
          type: string
          enum: [in_stock, out_of_stock]
        imageUrl: { type: string, nullable: true }
        category: { type: string, nullable: true }
        lastVerifiedAt: { type: string }
      required:
        - id
        - shopId
        - sellerId
        - name
        - price
        - quantity
        - colors
        - photos
        - sizes
        - categories
        - variations
        - reviews
        - totalSell
        - applyDiscount
        - tags
        - stockStatus
        - lastVerifiedAt
    ProductCreateInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        brand: { type: string, nullable: true }
        description: { type: string, nullable: true }
        price: { type: number, minimum: 0 }
        quantity: { type: number, minimum: 0 }
        colors:
          type: array
          items: { type: string }
        photos:
          type: array
          items: { type: string }
        sizes:
          type: array
          items: { type: string }
        categoryIds:
          type: array
          items: { type: string }
        weight: { type: number, nullable: true }
        dimension:
          oneOf:
            - { $ref: "#/components/schemas/Dimension" }
            - { type: "null" }
        variations:
          type: array
          items: { $ref: "#/components/schemas/ProductVariation" }
        tags:
          type: array
          items: { type: string }
        imageUrl: { type: string, nullable: true }
        category: { type: string, nullable: true }
        stockStatus:
          type: string
          enum: [in_stock, out_of_stock]
        applyDiscount: { type: boolean }
      required: [name]
    ProductUpdateInput:
      type: object
      properties:
        name: { type: string }
        brand: { type: string, nullable: true }
        description: { type: string, nullable: true }
        price: { type: number, minimum: 0 }
        quantity: { type: number, minimum: 0 }
        colors:
          type: array
          items: { type: string }
        photos:
          type: array
          items: { type: string }
        sizes:
          type: array
          items: { type: string }
        categoryIds:
          type: array
          items: { type: string }
        weight: { type: number, nullable: true }
        dimension:
          oneOf:
            - { $ref: "#/components/schemas/Dimension" }
            - { type: "null" }
        variations:
          type: array
          items: { $ref: "#/components/schemas/ProductVariation" }
        tags:
          type: array
          items: { type: string }
        imageUrl: { type: string, nullable: true }
        stockStatus:
          type: string
          enum: [in_stock, out_of_stock]
        applyDiscount: { type: boolean }
    AnalyzePhotoInput:
      type: object
      properties:
        imageBase64: { type: string }
      required: [imageBase64]
    ProductSuggestion:
      type: object
      properties:
        name: { type: string }
        category: { type: string }
        price: { type: number }
        tags:
          type: array
          items: { type: string }
      required: [name, category, price, tags]
...
    DeleteResult:
      type: object
      properties:
        success: { type: boolean }
      required: [success]
...
    ShopMember:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        email: { type: string, nullable: true }
        name: { type: string, nullable: true }
        role:
          type: string
          enum: [seller, sub_seller]
        createdAt: { type: string }
      required: [id, userId, role, createdAt]
    ShopInvitation:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        role:
          type: string
          enum: [seller, sub_seller]
        createdAt: { type: string }
      required: [id, email, role, createdAt]
    ShopMembersList:
      type: object
      properties:
        members:
          type: array
          items: { $ref: "#/components/schemas/ShopMember" }
        invitations:
          type: array
          items: { $ref: "#/components/schemas/ShopInvitation" }
      required: [members, invitations]
    InviteMemberInput:
      type: object
      properties:
        email: { type: string, format: email }
      required: [email]
...
    Discount:
      type: object
      properties:
        id: { type: string }
        productId: { type: string }
        code: { type: string, nullable: true }
        percentOff: { type: number }
        amountOff: { type: number }
        validFrom: { type: string }
        validTo: { type: string, nullable: true }
        isActive: { type: boolean }
      required: [id, productId, percentOff, amountOff, validFrom, isActive]
    DiscountCreateInput:
      type: object
      properties:
        code: { type: string, nullable: true }
        percentOff: { type: number, minimum: 0, maximum: 100 }
        amountOff: { type: number, minimum: 0 }
        validFrom: { type: string }
        validTo: { type: string, nullable: true }
        isActive: { type: boolean }