# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NearBuy Business API (MongoDB-backed)
 * OpenAPI spec version: 0.2.0
 */
import type { Category } from "./category";
import type { Dimension } from "./dimension";
import type { ProductStockStatus } from "./productStockStatus";
import type { ProductVariation } from "./productVariation";

export interface Product {
  id: string;
  shopId: string;
  sellerId: string;
  name: string;
  slug?: string | null;
  brand?: string | null;
  description?: string | null;
  price: number;
  quantity: number;
  colors: string[];
  photos: string[];
  sizes: string[];
  categories: Category[];
  weight?: number | null;
  dimension?: Dimension | null;
  variations: ProductVariation[];
  rating?: number | null;
  reviews: number;
  totalSell: number;
  applyDiscount: boolean;
  tags: string[];
  stockStatus: ProductStockStatus;
  imageUrl?: string | null;
  category?: string | null;
  lastVerifiedAt: string;
}