# lib/api-zod/src/generated/types/productSuggestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NearBuy Business API (MongoDB-backed)
 * OpenAPI spec version: 0.2.0
 */

export interface ProductSuggestion {
  name: string;
  category: string;
  price: number;
  tags: string[];
}